Updated readme and refactor the main files' names
