+++ README.md
@@ -1,20 +1,31 @@
 # PCE
 ## Run with python
-### Functions
-Most of the completed functions are located in the /LoadBalancing. It can solver CSP and Multi-commodities problem.
 
-Multi_Input_LoadBalancer is used to generated topology of the network.
-
-Multi_Input_Solver will generate the optimal solution based on minimizing the total cost.
-
-LB_Utilization_Solver will generate the optimal solution based on minimizing the total link capacity utilization.
-
-Connection.json is used as the query list and  located in the /test/data.
 ### Input
-The input is stored in the /test/data/connection.json. The format of the input connection is [Source node, Destination node, Bandwidth required, Latency required]
-For example, [1,10,15,20] means this connection query is to route traffic from Node 1 to Node 10, requring a bandwith of 10 and maximum latency of 20.
+The input is stored in the /test/data/connection.json. The format of the input connection is `[Source node, Destination node, Bandwidth required, Latency required]`
+For example, `[1,10,15,20]` means this connection query is to route traffic from Node 1 to Node 10, requring a bandwith of 10 and maximum latency of 20.
 
-For testing, a random connection generator is located at Utility/RandomConnectionGen.py. It can randomly generate one or multiple queries in one connection.json.+
+For testing, a random connection generator is located at `Utility/RandomConnectionGen.py`. It can randomly generate one or multiple queries in one connection.json.
+
+### Topology Generation
+Random connection generator is located at Utility/RandomConnectionGen.py. 
+```
+RandomConnectionGenerator(nodes, querynum, bw, latencylimit)
+```
+This will randomly generate one or multiple queries in one connection.json stored in `test/data/connection.json`. 
+
+### Output
+To get the optimal path and related objective value:
+For Minimizing the total path cost, go to `/LoadBalancing/SP_Solver.py`:
+```
+runSP_Solver()
+```
+
+For Minimizing the total path utilization, go to `/LoadBalancing/LB_Utilization_Solver.py`:
+```
+runLB_UT_Solver()
+```
+
 ### Environment Setup
 
 The network topology is generated by NetworkX and the traffic matrix computation is executed by Google OrTools Solver. 
@@ -29,7 +40,6 @@ Solver in this OR-Tools is used to solve for the optimal solution. It takes in t
 
 In this project, links are set to be binary variables which means if a link is selected, the corresponding variable is 1, if a link is not selected, the variable is 0.
 
-The object function will be the sum of cost of all selected links.
 
 
 ## Constrained Shortest Path (CSP)
