+++ test/test_load_balancing.py
@@ -1,6 +1,6 @@
 import unittest
 import json
-from LoadBalancing.Multi_Input_Solver import LB_Solver
+from LoadBalancing.SP_Solver import LB_Solver
 
 class Test_Load_Balancing_Solver(unittest.TestCase):
     def setUp(self):
